fix: update imports for moved ckpts
