+++ src/student_success_tool/preprocessing/checkpoints/pdp.py
@@ -3,7 +3,7 @@ import typing as t
 
 import pandas as pd
 
-from .. import utils
+from ... import utils
 
 LOGGER = logging.getLogger(__name__)
